+++ SpatialGDK/Source/SpatialGDK/Private/Interop/SpatialClassInfoManager.cpp
@@ -44,6 +44,15 @@ bool USpatialClassInfoManager::TryInit(USpatialNetDriver* InNetDriver)
 		return false;
 	}
 
+	if (SchemaDatabase->SchemaDatabaseVersion < ESchemaDatabaseVersion::LatestVersion)
+	{
+		UE_LOG(LogSpatialClassInfoManager, Error,
+			   TEXT("SchemaDatabase version old! Loaded: %d Expected: %d Please regenerate schema or turn off SpatialOS networking."),
+			   SchemaDatabase->SchemaDatabaseVersion, ESchemaDatabaseVersion::LatestVersion);
+		QuitGame();
+		return false;
+	}
+
 	return true;
 }