+++ SpatialGDK/Source/SpatialGDKEditor/Private/SchemaGenerator/SpatialGDKEditorSchemaGenerator.cpp
@@ -264,7 +264,7 @@ void GenerateSchemaFromClasses(const TArray<TSharedPtr<FUnrealType>>& TypeInfos,
 							   FComponentIdGenerator& IdGenerator)
 {
 	// Generate the actual schema.
-	FScopedSlowTask Progress((float)TypeInfos.Num(), LOCTEXT("GenerateSchemaFromClasses", "Generating Schema..."));
+	FScopedSlowTask Progress((float)TypeInfos.Num(), LOCTEXT("GenerateSchemaFromClasses", "Generating schema..."));
 	for (const auto& TypeInfo : TypeInfos)
 	{
 		Progress.EnterProgressFrame(1.f);
@@ -873,6 +873,8 @@ USchemaDatabase* InitialiseSchemaDatabase(const FString& PackagePath)
 	SchemaDatabase->ComponentSetIdToComponentIds.FindOrAdd(SpatialConstants::SERVER_AUTH_COMPONENT_SET_ID)
 		.ComponentIDs.Append(NetCullDistanceComponentIds);
 
+	SchemaDatabase->SchemaDatabaseVersion = ESchemaDatabaseVersion::LatestVersion;
+
 	return SchemaDatabase;
 }
 
@@ -933,7 +935,7 @@ bool SaveSchemaDatabase(USchemaDatabase* SchemaDatabase)
 		FString FullPath = FPaths::ConvertRelativePathToFull(FilePath);
 		FPaths::MakePlatformFilename(FullPath);
 		FMessageDialog::Debugf(FText::Format(
-			LOCTEXT("SchemaDatabaseLocked_Error", "Unable to save Schema Database to '{0}'! The file may be locked by another process."),
+			LOCTEXT("SchemaDatabaseLocked_Error", "Unable to save schema database to '{0}'! The file may be locked by another process."),
 			FText::FromString(FullPath)));
 		return false;
 	}
@@ -1051,10 +1053,11 @@ void CopyWellKnownSchemaFiles(const FString& GDKSchemaCopyDir, const FString& Co
 	}
 }
 
-bool RefreshSchemaFiles(const FString& SchemaOutputPath)
+bool RefreshSchemaFiles(const FString& SchemaOutputPath, const bool bDeleteExistingSchema /*= true*/,
+						const bool bCreateDirectoryTree /*= true*/)
 {
 	IPlatformFile& PlatformFile = FPlatformFileManager::Get().GetPlatformFile();
-	if (PlatformFile.DirectoryExists(*SchemaOutputPath))
+	if (bDeleteExistingSchema && PlatformFile.DirectoryExists(*SchemaOutputPath))
 	{
 		if (!PlatformFile.DeleteDirectoryRecursively(*SchemaOutputPath))
 		{
@@ -1065,7 +1068,7 @@ bool RefreshSchemaFiles(const FString& SchemaOutputPath)
 		}
 	}
 
-	if (!PlatformFile.CreateDirectoryTree(*SchemaOutputPath))
+	if (bCreateDirectoryTree && !PlatformFile.CreateDirectoryTree(*SchemaOutputPath))
 	{
 		UE_LOG(LogSpatialGDKSchemaGenerator, Error,
 			   TEXT("Could not create schema directory '%s'! Please make sure the parent directory is writeable."), *SchemaOutputPath);
@@ -1103,7 +1106,7 @@ bool LoadGeneratorStateFromSchemaDatabase(const FString& FileName)
 	{
 		FString AbsoluteFilePath = FPaths::ConvertRelativePathToFull(RelativeFileName);
 		UE_LOG(LogSpatialGDKSchemaGenerator, Error,
-			   TEXT("Schema Generation failed: Schema Database at %s is read only. Make it writable before generating schema"),
+			   TEXT("Schema generation failed: Schema Database at %s is read only. Make it writable before generating schema"),
 			   *AbsoluteFilePath);
 		return false;
 	}
@@ -1119,7 +1122,7 @@ bool LoadGeneratorStateFromSchemaDatabase(const FString& FileName)
 		if (SchemaDatabase == nullptr)
 		{
 			UE_LOG(LogSpatialGDKSchemaGenerator, Error,
-				   TEXT("Schema Generation failed: Failed to load existing schema database. If this continues, delete the schema database "
+				   TEXT("Schema generation failed: Failed to load existing schema database. If this continues, delete the schema database "
 						"and try again."));
 			return false;
 		}
@@ -1206,6 +1209,30 @@ bool GeneratedSchemaDatabaseExists()
 	IPlatformFile& PlatformFile = FPlatformFileManager::Get().GetPlatformFile();
 
 	return PlatformFile.FileExists(*RelativeSchemaDatabaseFilePath);
+}
+
+FSpatialGDKEditor::ESchemaDatabaseValidationResult ValidateSchemaDatabase()
+{
+	FFileStatData StatData = FPlatformFileManager::Get().GetPlatformFile().GetStatData(*RelativeSchemaDatabaseFilePath);
+	if (!StatData.bIsValid)
+	{
+		return FSpatialGDKEditor::NotFound;
+	}
+
+	const FString DatabaseAssetPath = FPaths::SetExtension(SpatialConstants::SCHEMA_DATABASE_ASSET_PATH, TEXT(".SchemaDatabase"));
+	const USchemaDatabase* const SchemaDatabase = Cast<USchemaDatabase>(FSoftObjectPath(DatabaseAssetPath).TryLoad());
+
+	if (SchemaDatabase == nullptr)
+	{
+		return FSpatialGDKEditor::NotFound;
+	}
+
+	if (SchemaDatabase->SchemaDatabaseVersion < ESchemaDatabaseVersion::LatestVersion)
+	{
+		return FSpatialGDKEditor::OldVersion;
+	}
+
+	return FSpatialGDKEditor::Ok;
 }
 
 void ResolveClassPathToSchemaName(const FString& ClassPath, const FString& SchemaName)
@@ -1344,11 +1371,11 @@ bool SpatialGDKGenerateSchema()
 
 	TArray<UObject*> AllClasses;
 	GetObjectsOfClass(UClass::StaticClass(), AllClasses);
-	if (!SpatialGDKGenerateSchemaForClasses(GetAllSupportedClasses(AllClasses),
-											GetDefault<USpatialGDKEditorSettings>()->GetGeneratedSchemaOutputFolder()))
+	if (!SpatialGDKGenerateSchemaForClasses(GetAllSupportedClasses(AllClasses)))
 	{
 		return false;
 	}
+	SpatialGDKSanitizeGeneratedSchema();
 
 	GenerateSchemaForSublevels();
 	GenerateSchemaForRPCEndpoints();
@@ -1445,6 +1472,45 @@ bool SpatialGDKGenerateSchemaForClasses(TSet<UClass*> Classes, FString SchemaOut
 	NextAvailableComponentId = IdGenerator.Peek();
 
 	return true;
+}
+
+template <class T>
+void SanitizeClassMap(TMap<FString, T>& Map, const TSet<FName>& ValidClassNames)
+{
+	for (auto Item = Map.CreateIterator(); Item; ++Item)
+	{
+		FString SanitizeName = Item->Key;
+		SanitizeName.RemoveFromEnd(TEXT("_C"));
+		if (!ValidClassNames.Contains(FName(*SanitizeName)))
+		{
+			UE_LOG(LogSpatialGDKSchemaGenerator, Log, TEXT("Found stale class (%s), removing from schema database."), *Item->Key);
+			Item.RemoveCurrent();
+		}
+	}
+}
+
+void SpatialGDKSanitizeGeneratedSchema()
+{
+	// Sanitize schema database, removing assets that no longer exist
+	FAssetRegistryModule& AssetRegistryModule = FModuleManager::LoadModuleChecked<FAssetRegistryModule>("AssetRegistry");
+
+	TArray<FAssetData> Assets;
+	AssetRegistryModule.Get().GetAllAssets(Assets, false);
+	TSet<FName> ValidClassNames;
+	for (const auto& Asset : Assets)
+	{
+		ValidClassNames.Add(FName(*Asset.ObjectPath.ToString()));
+	}
+
+	TArray<UObject*> AllClasses;
+	GetObjectsOfClass(UClass::StaticClass(), AllClasses);
+	for (const auto& SupportedClass : GetAllSupportedClasses(AllClasses))
+	{
+		ValidClassNames.Add(FName(*SupportedClass->GetPathName()));
+	}
+
+	SanitizeClassMap(ActorClassPathToSchema, ValidClassNames);
+	SanitizeClassMap(SubobjectClassPathToSchema, ValidClassNames);
 }
 
 } // namespace Schema