+++ SpatialGDK/Source/SpatialGDKFunctionalTests/SpatialGDK/SpatialComponentTest/SpatialComponentTestDummyComponent.h
@@ -1,0 +1,18 @@
+// Copyright (c) Improbable Worlds Ltd, All Rights Reserved
+
+#pragma once
+
+#include "Components/ActorComponent.h"
+#include "SpatialComponentTestDummyComponent.generated.h"
+
+/*
+ * Empty component to be added and removed from components so that the component callbacks can be tests
+ */
+UCLASS(NotBlueprintable, ClassGroup = SpatialFunctionalTest)
+class SPATIALGDKFUNCTIONALTESTS_API USpatialComponentTestDummyComponent : public USceneComponent
+{
+	GENERATED_BODY()
+
+public:
+	USpatialComponentTestDummyComponent();
+};